Normative: Mimic a prototype chain for static private fields
With this patch, reads and writes to a static private field which
is inherited from a superclass will behave similarly to reads and
writes to public properties: Reads will forward to the superclass
constructor, and writes will create a new value which is not
forwarded. Thanks to @rbuckton for the idea; this patch is an
iteration of spec drafts by him.